🔥 chore: type.ts 파일 삭제 ♻️ chore: 타입 폴더 하위에 저장 위치 및 경로 수정하기
