+++ @types/api/cards.d.ts
@@ -35,3 +35,19 @@ declare module '@planit-api' {
     imageUrl: string;
   };
 }
+
+// Todo Details Card
+export type TodoDetailsCardResponse = {
+  id: number;
+  title: string;
+  description: string;
+  tags: string[];
+  dueDate: string;
+  assignee: Assignee;
+  imageUrl: string;
+  teamId: string;
+  columnId: number;
+  dashboardId: number;
+  createdAt: string;
+  updatedAt: string;
+};
