+++ @types/api/commons.d.ts
@@ -2,4 +2,9 @@ declare module '@planit-api' {
   export type ErrorMessage = {
     message: string;
   };
+
+  export type ModalState = {
+    isOpen: boolean;
+    message: string;
+  };
 }
